refactor: 기존의 RabbitMQ 패키지 구조 변경
